logout, login

+++ taskboard/board/views.py
@@ -1,7 +1,8 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Task, TaskList, Group
-from .forms import TaskForm, TaskListForm, GroupInviteForm, EditProfileForm
-from django.contrib.auth import update_session_auth_hash
+from .forms import LoginForm, TaskForm, TaskListForm, GroupInviteForm, EditProfileForm
+from django.contrib.auth import login, authenticate
+from django.contrib.auth import logout
 from django.utils.timezone import now
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
@@ -10,23 +11,30 @@ from django.contrib.auth.decorators import login_required
 
 @login_required
 def task_list(request):
-    tasks = Task.objects.filter(user=request.user)
+    sort = request.GET.get('sort', 'title')
+    if sort not in ['title', 'due_date', 'is_completed']:
+        sort = 'title'
+    tasks = Task.objects.filter(user=request.user).order_by(sort)
     return render(request, 'board/task_list.html', {'tasks': tasks})
 
 def task_detail(request, pk):
     task = get_object_or_404(Task, pk=pk)
     return render(request, 'board/task_detail.html', {'task': task})
 
+@login_required
 def task_create(request):
     if request.method == "POST":
         form = TaskForm(request.POST)
         if form.is_valid():
-            form.save()
+            task = form.save(commit=False)
+            task.user = request.user
+            task.save()
             return redirect('task_list')
     else:
         form = TaskForm()
     return render(request, 'board/task_form.html', {'form': form})
 
+@login_required
 def task_edit(request, pk):
     task = get_object_or_404(Task, pk=pk)
     if request.method == "POST":
@@ -38,10 +46,13 @@ def task_edit(request, pk):
         form = TaskForm(instance=task)
     return render(request, 'board/task_form.html', {'form': form})
 
+@login_required
 def task_delete(request, pk):
-    task = get_object_or_404(Task, pk=pk)
-    task.delete()
-    return redirect('task_list')
+    task = get_object_or_404(Task, pk=pk, user=request.user)
+    if request.method == 'POST':
+        task.delete()
+        return redirect('board/task_list')
+    return render(request, 'board/task_confirm_delete.html', {'task': task})
 
 @login_required
 def profile(request):
@@ -49,13 +60,11 @@ def profile(request):
 
 @login_required
 def edit_profile(request):
+    user = request.user
     if request.method == 'POST':
-        form = EditProfileForm(request.POST, request.FILES, instance=request.user)
+        form = EditProfileForm(request.POST, instance=user)
         if form.is_valid():
             user = form.save(commit=False)
-            if form.cleaned_data['password']:
-                user.set_password(form.cleaned_data['password'])
-                update_session_auth_hash(request, user) 
             user.save() 
             return redirect('profile')
     else:
@@ -93,3 +102,40 @@ def register(request):
     else:
         form = UserCreationForm()
     return render(request, 'board/register.html', {'form': form})
+
+# def user_login(request):
+#     if request.method == 'POST':
+#         form = LoginForm(request, data=request.POST)
+#         if form.is_valid():
+#             username = form.cleaned_data['username']
+#             password = form.cleaned_data['password']
+#             user = authenticate(request, username=username, password=password)
+#             if user is not None:
+#                 login(request, user)
+#                 return redirect('task_list')  
+#             else:
+#                 form.add_error(None, "Invalid username or password")
+#     else:
+#         form = LoginForm()
+#     return render(request, 'login.html', {'form': form})
+
+def user_logout(request):
+    logout(request)
+    return redirect('task_list')
+
+def create_task_list(request):
+    if request.method == 'POST':
+        form = TaskListForm(request.POST)
+        if form.is_valid():
+            task_list = form.save(commit=False)
+            task_list.user = request.user  
+            task_list.save()
+            return redirect('task_list')
+    else:
+        form = TaskListForm()
+
+    return render(request, 'board/task_list_form.html', {'form': form})
+
+def calendar_view(request):
+    tasks = Task.objects.filter(user=request.user, event_date__gte=now()).order_by('event_date')
+    return render(request, 'board/calendar.html', {'tasks': tasks})
